refactor: move tpc power harnesses to core
