+++ benchbox/platforms/base/execution.py
@@ -37,15 +37,13 @@ from typing import Any
 from benchbox.core.constants import (
     GENERIC_POWER_DEFAULT_MEASUREMENT_ITERATIONS,
     GENERIC_POWER_DEFAULT_WARMUP_ITERATIONS,
-    TPCDS_POWER_DEFAULT_MEASUREMENT_ITERATIONS,
-    TPCDS_POWER_DEFAULT_WARMUP_ITERATIONS,
-    TPCH_POWER_DEFAULT_MEASUREMENT_ITERATIONS,
-    TPCH_POWER_DEFAULT_WARMUP_ITERATIONS,
 )
 from benchbox.core.errors import PlanCaptureError
 from benchbox.core.operations import OperationExecutor
+from benchbox.core.power_harnesses import resolve_power_harness
 from benchbox.core.results.builder import benchmark_family, normalize_benchmark_id
 from benchbox.core.results.models import QUERY_RUN_TYPE_MEASUREMENT, QUERY_RUN_TYPE_WARMUP
+from benchbox.core.tpch.platform_power import _power_query_result, _power_test_error_result
 from benchbox.platforms.base.connection_wrappers import (
     PlatformAdapterConnection,
     _make_stream_cursor,
@@ -53,56 +51,7 @@ from benchbox.platforms.base.connection_wrappers import (
 from benchbox.utils.dialect_utils import SQLTranslationError
 from benchbox.utils.printing import quiet_console
 
-
-def _power_query_result(
-    query_result: dict[str, Any],
-    *,
-    stream_id: int,
-    iteration: int,
-    run_type: str,
-) -> dict[str, Any]:
-    """Convert a TPC power query result to the adapter result shape."""
-    platform_result = {
-        "query_id": query_result["query_id"],
-        "execution_time_seconds": query_result["execution_time_seconds"],
-        "status": "SUCCESS" if query_result["success"] else "FAILED",
-        "rows_returned": query_result.get("result_count", 0),
-        "test_type": "power",
-        "stream_id": query_result.get("stream_id", stream_id),
-        "position": query_result.get("position", 0),
-        "iteration": iteration,
-        "run_type": run_type,
-    }
-    # Gate-only value-digest oracle: forward the full-result digest when the power
-    # driver emitted one (behind BENCHBOX_EMIT_RESULT_DIGEST). Additive — absent on
-    # a normal run, so the payload shape is unchanged.
-    if query_result.get("result_digest") is not None:
-        platform_result["result_digest"] = query_result["result_digest"]
-    if not query_result["success"]:
-        platform_result["error"] = query_result.get("error", "Unknown error")
-    return platform_result
-
-
-def _power_test_error_result(
-    error: str,
-    *,
-    iteration: int | None = None,
-    run_type: str | None = None,
-) -> dict[str, Any]:
-    """Build a failed power-test sentinel result."""
-    result: dict[str, Any] = {
-        "query_id": "power_test_error",
-        "execution_time_seconds": 0.0,
-        "status": "FAILED",
-        "rows_returned": 0,
-        "error": error,
-        "test_type": "power",
-    }
-    if iteration is not None:
-        result["iteration"] = iteration
-    if run_type is not None:
-        result["run_type"] = run_type
-    return result
+__all__ = ["TestDriversMixin", "_power_query_result", "_power_test_error_result"]
 
 
 class TestDriversMixin:
@@ -121,153 +70,32 @@ class TestDriversMixin:
       `_build_query_failure_result`, `_build_dry_run_result`
     """
 
+    def _make_power_connection_adapter(self, connection: Any, benchmark_id: str, scale_factor: float):
+        """Wrap a platform connection for core TPC power harnesses."""
+        connection_adapter = PlatformAdapterConnection(_make_stream_cursor(connection), self)
+        connection_adapter.benchmark_type = benchmark_id
+        connection_adapter.scale_factor = scale_factor
+        return connection_adapter
+
+    def _make_direct_power_connection_adapter(self, connection: Any, benchmark_id: str, scale_factor: float):
+        """Wrap a direct platform connection for single-stream core TPC power harnesses."""
+        connection_adapter = PlatformAdapterConnection(connection, self)
+        connection_adapter.benchmark_type = benchmark_id
+        connection_adapter.scale_factor = scale_factor
+        return connection_adapter
+
     def _execute_tpch_power_test(self, benchmark, connection: Any, run_config: dict) -> list[dict[str, Any]]:
-        """Execute TPC-H Power Test using production TPCHPowerTest implementation."""
-        from benchbox.core.tpch.power_test import TPCHPowerTest
+        """Execute TPC-H Power Test using the core TPCH power harness."""
+        from benchbox.core.tpch.platform_power import execute_tpch_power_test
 
-        console = quiet_console
-
-        try:
-            # Extract configuration
-            scale_factor = run_config.get("scale_factor", 1.0)
-            seed = run_config.get("seed")
-            validation_mode = run_config.get("validation_mode")  # Universal validation mode
-            stream_id = run_config.get("stream_id", 0)
-            query_subset = run_config.get("query_subset")
-            getattr(self, "get_target_dialect", lambda: "standard")()
-            verbose = run_config.get("verbose", False)
-            timeout = run_config.get("timeout")
-            iterations = run_config.get("iterations", TPCH_POWER_DEFAULT_MEASUREMENT_ITERATIONS)
-            warm_up_iterations = run_config.get("warm_up_iterations", TPCH_POWER_DEFAULT_WARMUP_ITERATIONS)
-            fail_fast = run_config.get("power_fail_fast", False)
-
-            console.print(
-                f"[green]Running TPC-H Power Test (Scale Factor: {scale_factor}, Stream ID: {stream_id})[/green]"
-            )
-            console.print(f"[green]Warm-up runs: {warm_up_iterations}, Measurement runs: {iterations}[/green]")
-
-            # Create connection adapter that wraps the platform adapter connection
-            connection_adapter = PlatformAdapterConnection(connection, self)
-            # Configure benchmark context for query validation
-            connection_adapter.benchmark_type = "tpch"
-            connection_adapter.scale_factor = scale_factor
-
-            all_results = []
-
-            # Warm-up runs
-            for i in range(warm_up_iterations):
-                current_stream_id = i  # Start at 0 for warmup
-                console.print(f"[cyan]--- Warm-up Run {i + 1}/{warm_up_iterations} ---[/cyan]")
-                power_test = TPCHPowerTest(
-                    benchmark=benchmark,
-                    connection=connection_adapter,
-                    scale_factor=scale_factor,
-                    seed=seed,
-                    stream_id=current_stream_id,
-                    verbose=verbose,
-                    timeout=timeout,
-                    dialect=self.get_target_dialect(),
-                    validation_mode=validation_mode,
-                    query_subset=query_subset,
-                )
-                # Mirror the power test's resolved validation policy instead of
-                # re-deriving it from stream IDs in the adapter.
-                connection_adapter._validate_row_count = power_test.config.validation
-                power_test_result = power_test.run()
-                for query_result in power_test_result.query_results:
-                    all_results.append(
-                        _power_query_result(
-                            query_result,
-                            stream_id=current_stream_id,
-                            iteration=0,
-                            run_type="warmup",
-                        )
-                    )
-
-            # Measurement runs
-            for i in range(iterations):
-                current_stream_id = warm_up_iterations + i  # Continue from where warmup left off
-                console.print(f"[cyan]--- Measurement Run {i + 1}/{iterations} ---[/cyan]")
-                power_test = TPCHPowerTest(
-                    benchmark=benchmark,
-                    connection=connection_adapter,
-                    scale_factor=scale_factor,
-                    seed=seed,
-                    stream_id=current_stream_id,
-                    verbose=verbose,
-                    timeout=timeout,
-                    dialect=self.get_target_dialect(),
-                    validation_mode=validation_mode,
-                    query_subset=query_subset,
-                )
-                # Mirror the power test's resolved validation policy instead of
-                # re-deriving it from stream IDs in the adapter.
-                connection_adapter._validate_row_count = power_test.config.validation
-
-                # Execute the power test
-                power_test_result = power_test.run()
-
-                # Cache the power test result for metric extraction
-                self._last_power_test_result = power_test_result
-
-                # Display results
-                if power_test_result.success:
-                    success_rate = power_test_result.queries_successful / max(power_test_result.queries_executed, 1)
-                    console.print(
-                        f"[green]✅ TPC-H Power Test completed: Power@Size = {power_test_result.power_at_size:.2f}[/green]"
-                    )
-                    console.print(
-                        f"  Queries executed: {power_test_result.queries_executed}, Successful: {power_test_result.queries_successful}"
-                    )
-                    console.print(f"  Success rate: {success_rate:.1%} (TPC-H requires ≥95%)")
-                    console.print(f"  Total execution time: {power_test_result.total_time:.2f}s")
-                else:
-                    console.print("[red]❌ TPC-H Power Test failed[/red]")
-                    for error in power_test_result.errors:
-                        console.print(f"  Error: {error}")
-
-                # Convert TPCHPowerTestResult to platform adapter format
-                query_results = []
-                for query_result in power_test_result.query_results:
-                    query_results.append(
-                        _power_query_result(
-                            query_result,
-                            stream_id=current_stream_id,
-                            iteration=i + 1,
-                            run_type="measurement",
-                        )
-                    )
-                all_results.extend(query_results)
-
-                # Abort remaining iterations if all queries failed (connection/infra issue)
-                # or if fail_fast is enabled and any query failed
-                if not power_test_result.success:
-                    if power_test_result.queries_successful == 0:
-                        console.print("[yellow]⚠️  All queries failed - aborting remaining measurement runs[/yellow]")
-                        # If the factory itself failed (no queries ran at all), add a sentinel
-                        # so total_queries > 0 and the benchmark is marked FAILED, not PASSED.
-                        if not query_results:
-                            all_results.append(
-                                _power_test_error_result(
-                                    "; ".join(power_test_result.errors)
-                                    if power_test_result.errors
-                                    else "Power test failed",
-                                    iteration=i + 1,
-                                    run_type="measurement",
-                                )
-                            )
-                        break
-                    if fail_fast:
-                        console.print(
-                            "[yellow]⚠️  Query failures detected (fail_fast enabled) - aborting remaining runs[/yellow]"
-                        )
-                        break
-
-            return all_results
-
-        except Exception as e:
-            console.print(f"[red]❌ TPC-H Power Test failed: {e}[/red]")
-            return [_power_test_error_result(str(e))]
+        return execute_tpch_power_test(
+            self,
+            benchmark,
+            connection,
+            run_config,
+            make_connection_adapter=self._make_direct_power_connection_adapter,
+            console=quiet_console,
+        )
 
     def _execute_generic_power_test(self, benchmark, connection: Any, run_config: dict) -> list[dict[str, Any]]:
         """Execute power test for non-TPC benchmarks with warmup + iterations.
@@ -355,169 +183,17 @@ class TestDriversMixin:
         return all_measurement_results
 
     def _execute_tpcds_power_test(self, benchmark, connection: Any, run_config: dict) -> list[dict[str, Any]]:
-        """Execute TPC-DS Power Test using production TPCDSPowerTest implementation."""
-        from benchbox.core.expected_results.tpcds_results import set_config_validation_mode
-        from benchbox.core.tpcds.power_test import TPCDSPowerTest
+        """Execute TPC-DS Power Test using the core TPCDS power harness."""
+        from benchbox.core.tpcds.platform_power import execute_tpcds_power_test
 
-        console = quiet_console
-
-        try:
-            # Extract configuration
-            scale_factor = run_config.get("scale_factor", 1.0)
-            seed = run_config.get("seed", 1)
-            validation_mode = run_config.get("validation_mode")  # Universal validation mode
-            stream_id = run_config.get("stream_id", 0)
-            query_subset = run_config.get("query_subset")
-            dialect = getattr(self, "get_target_dialect", lambda: "standard")()
-            verbose = run_config.get("verbose", False)
-            timeout = run_config.get("timeout")
-            iterations = run_config.get("iterations", TPCDS_POWER_DEFAULT_MEASUREMENT_ITERATIONS)
-            warm_up_iterations = run_config.get("warm_up_iterations", TPCDS_POWER_DEFAULT_WARMUP_ITERATIONS)
-            fail_fast = run_config.get("power_fail_fast", False)
-
-            # Set TPC-DS validation mode from config (takes precedence over environment variable)
-            set_config_validation_mode(validation_mode)
-
-            console.print(
-                f"[green]Running TPC-DS Power Test (Scale Factor: {scale_factor}, Stream ID: {stream_id})[/green]"
-            )
-            console.print(f"[green]Warm-up runs: {warm_up_iterations}, Measurement runs: {iterations}[/green]")
-
-            # Create connection factory that wraps the platform adapter connection
-            def connection_factory():
-                # Create thread-safe cursor for concurrent stream execution
-                # See: https://duckdb.org/docs/stable/guides/python/multiple_threads
-                # Some clients (e.g. ClickHouseLocalClient) expose execute() directly
-                # without a cursor() method.  In that case we wrap the connection in a
-                # non-closing proxy so that PlatformAdapterConnection.close() (called at
-                # the end of every power-test run) doesn't destroy the shared underlying
-                # connection before the next warmup/measurement run can use it.
-                stream_cursor = _make_stream_cursor(connection)
-
-                conn_wrapper = PlatformAdapterConnection(stream_cursor, self)
-                # Configure benchmark context for query validation
-                conn_wrapper.benchmark_type = "tpcds"
-                conn_wrapper.scale_factor = scale_factor
-                return conn_wrapper
-
-            all_results = []
-
-            # Warm-up runs
-            for i in range(warm_up_iterations):
-                current_stream_id = i  # Start at 0 for warmup
-                console.print(f"[cyan]--- Warm-up Run {i + 1}/{warm_up_iterations} ---[/cyan]")
-                power_test = TPCDSPowerTest(
-                    benchmark=benchmark,
-                    connection_factory=connection_factory,
-                    scale_factor=scale_factor,
-                    seed=seed,
-                    stream_id=current_stream_id,
-                    verbose=verbose,
-                    timeout=timeout,
-                    dialect=self.get_target_dialect(),
-                    query_subset=query_subset,
-                )
-                power_test_result = power_test.run()
-                for query_result in power_test_result.query_results:
-                    all_results.append(
-                        _power_query_result(
-                            query_result,
-                            stream_id=current_stream_id,
-                            iteration=0,
-                            run_type="warmup",
-                        )
-                    )
-
-            # Measurement runs
-            for i in range(iterations):
-                current_stream_id = warm_up_iterations + i  # Continue from where warmup left off
-                console.print(f"[cyan]--- Measurement Run {i + 1}/{iterations} ---[/cyan]")
-                power_test = TPCDSPowerTest(
-                    benchmark=benchmark,
-                    connection_factory=connection_factory,
-                    scale_factor=scale_factor,
-                    seed=seed,
-                    stream_id=current_stream_id,
-                    verbose=verbose,
-                    timeout=timeout,
-                    dialect=self.get_target_dialect(),
-                    query_subset=query_subset,
-                )
-
-                # Execute the power test
-                power_test_result = power_test.run()
-
-                # Cache the power test result for metric extraction
-                self._last_power_test_result = power_test_result
-
-                # Display results
-                if self.very_verbose:
-                    with contextlib.suppress(Exception):
-                        console.print(f"[dim]Target dialect: {dialect} | Detailed per-query results:[/dim]")
-                    for qr in power_test_result.query_results:
-                        qname = f"q{qr.get('query_id')}"
-                        dur = qr.get("execution_time_seconds", 0.0)
-                        status = "SUCCESS" if qr.get("success") else "FAILED"
-                        rows = qr.get("result_count", 0)
-                        console.print(f"  • {qname}: {dur:.2f}s, {status}, rows={rows}")
-
-                if power_test_result.success:
-                    success_rate = power_test_result.queries_successful / max(power_test_result.queries_executed, 1)
-                    console.print(
-                        f"[green]✅ TPC-DS Power Test completed: Power@Size = {power_test_result.power_at_size:.2f}[/green]"
-                    )
-                    console.print(
-                        f"  Queries executed: {power_test_result.queries_executed}, Successful: {power_test_result.queries_successful}"
-                    )
-                    console.print(f"  Success rate: {success_rate:.1%}")
-                    console.print(f"  Total execution time: {power_test_result.total_time:.2f}s")
-                else:
-                    console.print("[red]❌ TPC-DS Power Test failed[/red]")
-                    for error in power_test_result.errors:
-                        console.print(f"  Error: {error}")
-
-                # Convert TPCDSPowerTestResult to platform adapter format
-                query_results = []
-                for query_result in power_test_result.query_results:
-                    query_results.append(
-                        _power_query_result(
-                            query_result,
-                            stream_id=current_stream_id,
-                            iteration=i + 1,
-                            run_type="measurement",
-                        )
-                    )
-                all_results.extend(query_results)  # Accumulate results from all iterations
-
-                # Abort remaining iterations if all queries failed (connection/infra issue)
-                # or if fail_fast is enabled and any query failed
-                if not power_test_result.success:
-                    if power_test_result.queries_successful == 0:
-                        console.print("[yellow]⚠️  All queries failed - aborting remaining measurement runs[/yellow]")
-                        # If the factory itself failed (no queries ran at all), add a sentinel
-                        # so total_queries > 0 and the benchmark is marked FAILED, not PASSED.
-                        if not query_results:
-                            all_results.append(
-                                _power_test_error_result(
-                                    "; ".join(power_test_result.errors)
-                                    if power_test_result.errors
-                                    else "Power test failed",
-                                    iteration=i + 1,
-                                    run_type="measurement",
-                                )
-                            )
-                        break
-                    if fail_fast:
-                        console.print(
-                            "[yellow]⚠️  Query failures detected (fail_fast enabled) - aborting remaining runs[/yellow]"
-                        )
-                        break
-
-            return all_results
-
-        except Exception as e:
-            console.print(f"[red]❌ TPC-DS Power Test failed: {e}[/red]")
-            return [_power_test_error_result(str(e))]
+        return execute_tpcds_power_test(
+            self,
+            benchmark,
+            connection,
+            run_config,
+            make_connection_adapter=self._make_power_connection_adapter,
+            console=quiet_console,
+        )
 
     def _execute_tpcds_throughput_test(self, benchmark, connection: Any, run_config: dict) -> list[dict[str, Any]]:
         """Execute TPC-DS Throughput Test using production TPCDSThroughputTest implementation."""
@@ -1013,13 +689,10 @@ class TestDriversMixin:
         benchmark_name = self._resolve_benchmark_slug(benchmark, run_config)
         benchmark_id = normalize_benchmark_id(benchmark_name)
 
-        if benchmark_id == "tpch":
-            return self._execute_tpch_power_test(benchmark, connection, run_config)
-        elif benchmark_id == "tpcds":
-            return self._execute_tpcds_power_test(benchmark, connection, run_config)
-        else:
-            # Use generic power test handler for non-TPC benchmarks
-            return self._execute_generic_power_test(benchmark, connection, run_config)
+        harness = resolve_power_harness(benchmark_id)
+        if harness is not None:
+            return getattr(self, harness.adapter_method)(benchmark, connection, run_config)
+        return self._execute_generic_power_test(benchmark, connection, run_config)
 
     def _execute_throughput_test(self, benchmark, connection: Any, run_config: dict) -> list[dict[str, Any]]:
         """Execute TPC Throughput Test with concurrent query streams.
